Did Some stuff. Now its super slow
