+++ app/build.gradle
@@ -24,9 +24,9 @@ dependencies {
     androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
+    compile files('libs/TangoSDK_Gankino_Java.jar')
     compile 'com.android.support:appcompat-v7:25.1.1'
     compile 'com.android.support.constraint:constraint-layout:1.0.2'
     compile 'com.android.support:design:25.1.1'
     testCompile 'junit:junit:4.12'
-    compile files('libs/TangoSDK_Gankino_Java.jar')
 }
