+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,5 @@
 <resources>
     <string name="app_name">MotionTrackingInClass</string>
     <string name="action_settings">Settings</string>
+    <string name="title_activity_load_adf">LoadADFActivity</string>
 </resources>
